third_party/blink/web_tests/fast/loader/hashchange-event-expected.txt

ALERT: Test number 1.
hashchange event received by window-event-listener.
A total of 1 events have been received.
ALERT: Test number 1.
hashchange event received by body-onhashchange-attribute.
A total of 2 events have been received.
ALERT: Test number 1.
hashchange event received by window-event-listener.
A total of 3 events have been received.
ALERT: Test number 1.
hashchange event received by body-onhashchange-attribute.
A total of 4 events have been received.
ALERT: Test number 1.
hashchange event received by window-event-listener.
A total of 5 events have been received.
ALERT: Test number 1.
hashchange event received by body-onhashchange-attribute.
A total of 6 events have been received.
ALERT: Test number 1.
hashchange event received by window-event-listener.
A total of 7 events have been received.
ALERT: Test number 1.
hashchange event received by body-onhashchange-attribute.
A total of 8 events have been received.
ALERT: Test number 2.
hashchange event received by window-event-listener.
A total of 1 events have been received.
ALERT: Test number 2.
hashchange event received by body-onhashchange-attribute.
A total of 2 events have been received.
ALERT: Test number 2.
hashchange event received by window-event-listener.
A total of 3 events have been received.
ALERT: Test number 2.
hashchange event received by body-onhashchange-attribute.
A total of 4 events have been received.
ALERT: Test number 2.
hashchange event received by window-event-listener.
A total of 5 events have been received.
ALERT: Test number 2.
hashchange event received by body-onhashchange-attribute.
A total of 6 events have been received.
ALERT: Test number 2.
hashchange event received by window-event-listener.
A total of 7 events have been received.
ALERT: Test number 2.
hashchange event received by body-onhashchange-attribute.
A total of 8 events have been received.
ALERT: Test number 2.
hashchange event received by window-event-listener.
A total of 9 events have been received.
ALERT: Test number 2.
hashchange event received by body-onhashchange-attribute.
A total of 10 events have been received.
This test checks to make sure the hashchange event is fired when the value of location.hash changes.
This includes testing the transition from no fragment identifier to an empty fragment identifier *not* generating an event, as these cases are equivalent as far as location.hash is concerned.
Clicking the links below manually should also result in the event being fired when the hash actually changes.
Go to empty hash
Go to non-empty hash
Empty hash anchor
Non-empty hash anchor
